Parse go.mod and compare if present
Add the ability to parse a go.mod file when vendor.conf is absent.
dependencies found in a go.mod can be compared with vendor.conf
As a result of how go modules formats commit hashes, I also trimmed the commit hashes from vendor.conf for comparison. I think it cleans up the resulting release file, and is the best solution for go.mod's short hashes.

Signed-off-by: Peter Hunt <pehunt@redhat.com>